move condition tests to separate file
